+++ package.json
@@ -14,6 +14,7 @@
     "concurrently": "^9.0.1",
     "cors": "^2.8.5",
     "express": "^4.21.1",
+    "js-yaml": "^4.1.0",
     "mongoose": "^8.8.1",
     "swagger-jsdoc": "^6.2.8",
     "swagger-ui-express": "^5.0.1",
@@ -24,6 +25,7 @@
     "@types/concurrently": "^7.0.0",
     "@types/cors": "^2.8.17",
     "@types/express": "^5.0.0",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.7.5",
     "@types/swagger-jsdoc": "^6.0.4",
     "@types/swagger-ui-express": "^4.1.8",
