+++ src/routes/mdb-post.ts
@@ -8,48 +8,6 @@ interface MongoError extends Error {
   keyValue?: Record<string, any>;
 }
 
-/**
- * @openapi
- * /new-user:
- *   post:
- *     tags:
- *        - Create New User
- *     description: Create a new user
- *     requestBody:
- *        required: true
- *        content:
- *          application/json:
- *            schema:
- *              type: object
- *              properties:
- *                 userid:
- *                   type: string
- *                 first_name:
- *                   type: string
- *                 last_name:
- *                   type: string
- *                 courses:
- *                   type: array
- *                   items:
- *                     type: string
- *                 enrolled_courses:
- *                   type: array
- *                   items:
- *                     type: string
- *              required:
- *                - userid
- *                - first_name
- *                - last_name
- *                - courses
- *                - enrolled_courses
- *     responses:
- *        201:
- *          description: New user created successfully
- *        400:
- *          description: Malformed syntax or invalid data
- *        409:
- *          description: User validation failed 
- */
 router.post("/new-user", async (req: Request, res: Response) => {
   try {
     const pl = {
@@ -60,10 +18,11 @@ router.post("/new-user", async (req: Request, res: Response) => {
     const result = await newUser.save();
     res.status(201).json(result);
   } catch (error) {
-    if((error as MongoError).code === 11000) {
-      res.status(409).json({ 
-        message: `${(error as MongoError).keyValue?.userid} already exist!`
-      })
+    if ((error as MongoError).code === 11000) {
+      res.status(409).json({
+        message: `${(error as MongoError).keyValue?.userid} already exist!`,
+      });
+      return;
     }
     res.status(400).json({ message: `${(error as MongoError).message}` });
   }
