+++ src/utils/index.ts
@@ -1,0 +1,9 @@
+export {
+  getQuestions,
+  fetchQuestions,
+  validateWeek,
+  extractor,
+  getActiveDates,
+} from "./helperFunctions";
+
+export { GameData, CourseEnrolled } from "./types";
