+++ src/utils/types.ts
@@ -1,0 +1,10 @@
+export type GameData = {
+  [key: string]:
+    | {
+        [key: string]: string;
+      }
+    | string
+    | {};
+};
+
+export type CourseEnrolled = "GES" | "GES2";
